requirements.txt
cryptography
Flask
Pillow
